make casting throw if overflow

+++ server/src/test/java/org/elasticsearch/cluster/ClusterStateTests.java
@@ -1242,6 +1242,6 @@ public class ClusterStateTests extends ESTestCase {
             }
         }
 
-        return (int) chunkCount;
+        return Math.toIntExact(chunkCount);
     }
 }
